refactor #36 메인페이지 코드 리팩토링
