+++ src/components/homeMain/atoms/MainShoppingSwiper.vue
